switch new etcd gathering
