--- pkg/operator/configobservation/etcd/OWNERS
@@ -1,8 +1,0 @@
-reviewers:
-  - ironcladlou
-  - hexfusion
-  - retroflexer
-approvers:
-  - ironcladlou
-  - hexfusion
-  - retroflexer
